# Issue #45 Speed up XlsxReader file parsing.