This PR speeds up the `.xlsx` file parsing in `XlsxReader.py`. Here are my measurements of the speedup, using the release 2.1.3 dataset. 

Speed of the original implementation using the validation dataset (which is a four project list)

Original implementation, serial execution: **27 seconds**
New implementation, serial execution: **13 seconds**
Original implementation, parallel execution: **9 seconds**
New implementation, parallel execution: **5 seconds**

Test it using the command appropriate for your environment. With something like

```
python main.py -I /Users/akey/ProprietaryData/inputs -o /Users/akey/ProprietaryData/outputs --validate
```


Or, use the following procedure to duplicate the timing results.

The times were measured running a validation operation with the following command (adjust it to whatever file paths work on your system)

```
date && python main.py -i /Users/akey/ProprietaryData/inputs -o /Users/akey/ProprietaryData/outputs --validate && date
```

The `date &&` and `&& date` at the beginning and end display the start and end timestamps of the run.

Making a standard log of start and end times could be added with another commit.